@@ -11,14 +11,14 @@ var __metadata = (this && this.__metadata) || function (k, v) {
 var __param = (this && this.__param) || function (paramIndex, decorator) {
     return function (target, key) { decorator(target, key, paramIndex); }
 };
-var core_1 = require('@angular/core');
-var map_1 = require('../map/map');
-var group_1 = require('../group/group');
-var map_service_1 = require('../services/map.service');
-var group_service_1 = require('../services/group.service');
-var popup_service_1 = require('../services/popup.service');
-var globalId_service_1 = require('../services/globalId.service');
-var helper_service_1 = require('../services/helper.service');
+var core_1 = require("@angular/core");
+var map_1 = require("../map/map");
+var group_1 = require("../group/group");
+var map_service_1 = require("../services/map.service");
+var group_service_1 = require("../services/group.service");
+var popup_service_1 = require("../services/popup.service");
+var globalId_service_1 = require("../services/globalId.service");
+var helper_service_1 = require("../services/helper.service");
 var Lealflet = require('leaflet');
 var GeoJsonElement = (function () {
     function GeoJsonElement(mapService, groupService, popupService, guidService, helperService, LeafletElement, LeafletGroup) {
@@ -51,22 +51,28 @@ var GeoJsonElement = (function () {
     GeoJsonElement.prototype.ngDoCheck = function () {
         var map = this.mapService.getMap();
     };
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', Object)
-    ], GeoJsonElement.prototype, "geojson", void 0);
-    GeoJsonElement = __decorate([
-        core_1.Component({
-            moduleId: module.id,
-            selector: 'geojson-element',
-            templateUrl: 'geojson.html',
-            styleUrls: ['geojson.css']
-        }),
-        __param(5, core_1.Optional()),
-        __param(6, core_1.Optional()), 
-        __metadata('design:paramtypes', [map_service_1.MapService, group_service_1.GroupService, popup_service_1.PopupService, globalId_service_1.GuidService, helper_service_1.HelperService, map_1.LeafletElement, group_1.LeafletGroup])
-    ], GeoJsonElement);
     return GeoJsonElement;
 }());
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", Object)
+], GeoJsonElement.prototype, "geojson", void 0);
+GeoJsonElement = __decorate([
+    core_1.Component({
+        moduleId: module.id,
+        selector: 'geojson-element',
+        templateUrl: 'geojson.html',
+        styleUrls: ['geojson.css']
+    }),
+    __param(5, core_1.Optional()),
+    __param(6, core_1.Optional()),
+    __metadata("design:paramtypes", [map_service_1.MapService,
+        group_service_1.GroupService,
+        popup_service_1.PopupService,
+        globalId_service_1.GuidService,
+        helper_service_1.HelperService,
+        map_1.LeafletElement,
+        group_1.LeafletGroup])
+], GeoJsonElement);
 exports.GeoJsonElement = GeoJsonElement;
 //# sourceMappingURL=geojson.js.map
