@@ -8,9 +8,8 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
 var __metadata = (this && this.__metadata) || function (k, v) {
     if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
 };
-var core_1 = require('@angular/core');
-var map_service_1 = require('../services/map.service');
-var Lealflet = require('leaflet');
+var core_1 = require("@angular/core");
+var map_service_1 = require("../services/map.service");
 var LeafletElement = (function () {
     function LeafletElement(mapService) {
         this.mapService = mapService;
@@ -68,57 +67,57 @@ var LeafletElement = (function () {
             this.layerControlObject = L.control.layers(this.mapService.getBasemaps(), this.mapService.getOverlays()).addTo(map);
         }
     };
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', Number)
-    ], LeafletElement.prototype, "lat", void 0);
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', Number)
-    ], LeafletElement.prototype, "lon", void 0);
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', Number)
-    ], LeafletElement.prototype, "x", void 0);
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', Number)
-    ], LeafletElement.prototype, "y", void 0);
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', Number)
-    ], LeafletElement.prototype, "zoom", void 0);
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', Number)
-    ], LeafletElement.prototype, "minZoom", void 0);
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', Number)
-    ], LeafletElement.prototype, "maxZoom", void 0);
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', Boolean)
-    ], LeafletElement.prototype, "layerControl", void 0);
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', Object)
-    ], LeafletElement.prototype, "crs", void 0);
-    __decorate([
-        core_1.ViewChild('map'), 
-        __metadata('design:type', core_1.ElementRef)
-    ], LeafletElement.prototype, "mapElement", void 0);
-    LeafletElement = __decorate([
-        core_1.Component({
-            moduleId: module.id,
-            selector: 'leaf-element',
-            templateUrl: 'map.html',
-            styleUrls: ['map.css'],
-            providers: [map_service_1.MapService]
-        }), 
-        __metadata('design:paramtypes', [map_service_1.MapService])
-    ], LeafletElement);
     return LeafletElement;
 }());
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", Number)
+], LeafletElement.prototype, "lat", void 0);
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", Number)
+], LeafletElement.prototype, "lon", void 0);
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", Number)
+], LeafletElement.prototype, "x", void 0);
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", Number)
+], LeafletElement.prototype, "y", void 0);
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", Number)
+], LeafletElement.prototype, "zoom", void 0);
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", Number)
+], LeafletElement.prototype, "minZoom", void 0);
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", Number)
+], LeafletElement.prototype, "maxZoom", void 0);
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", Boolean)
+], LeafletElement.prototype, "layerControl", void 0);
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", Object)
+], LeafletElement.prototype, "crs", void 0);
+__decorate([
+    core_1.ViewChild('map'),
+    __metadata("design:type", core_1.ElementRef)
+], LeafletElement.prototype, "mapElement", void 0);
+LeafletElement = __decorate([
+    core_1.Component({
+        moduleId: module.id.toString(),
+        selector: 'leaf-element',
+        templateUrl: 'map.html',
+        styleUrls: ['map.css'],
+        providers: [map_service_1.MapService]
+    }),
+    __metadata("design:paramtypes", [map_service_1.MapService])
+], LeafletElement);
 exports.LeafletElement = LeafletElement;
 //# sourceMappingURL=map.js.map
