@@ -11,15 +11,15 @@ var __metadata = (this && this.__metadata) || function (k, v) {
 var __param = (this && this.__param) || function (paramIndex, decorator) {
     return function (target, key) { decorator(target, key, paramIndex); }
 };
-var core_1 = require('@angular/core');
-var map_1 = require('../map/map');
-var group_1 = require('../group/group');
-var map_service_1 = require('../services/map.service');
-var group_service_1 = require('../services/group.service');
-var popup_service_1 = require('../services/popup.service');
-var globalId_service_1 = require('../services/globalId.service');
-var helper_service_1 = require('../services/helper.service');
-var path_1 = require('../models/path');
+var core_1 = require("@angular/core");
+var map_1 = require("../map/map");
+var group_1 = require("../group/group");
+var map_service_1 = require("../services/map.service");
+var group_service_1 = require("../services/group.service");
+var popup_service_1 = require("../services/popup.service");
+var globalId_service_1 = require("../services/globalId.service");
+var helper_service_1 = require("../services/helper.service");
+var path_1 = require("../models/path");
 var Lealflet = require('leaflet');
 var PolygonElement = (function () {
     function PolygonElement(mapService, groupService, popupService, guidService, helperService, LeafletElement, LeafletGroup) {
@@ -73,34 +73,40 @@ var PolygonElement = (function () {
             }
         }
     };
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', Array)
-    ], PolygonElement.prototype, "latlngs", void 0);
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', Object)
-    ], PolygonElement.prototype, "Options", void 0);
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', String)
-    ], PolygonElement.prototype, "mouseover", void 0);
-    __decorate([
-        core_1.Input(), 
-        __metadata('design:type', String)
-    ], PolygonElement.prototype, "onclick", void 0);
-    PolygonElement = __decorate([
-        core_1.Component({
-            moduleId: module.id,
-            selector: 'polygon-element',
-            templateUrl: 'polygon.html',
-            styleUrls: ['polygon.css']
-        }),
-        __param(5, core_1.Optional()),
-        __param(6, core_1.Optional()), 
-        __metadata('design:paramtypes', [map_service_1.MapService, group_service_1.GroupService, popup_service_1.PopupService, globalId_service_1.GuidService, helper_service_1.HelperService, map_1.LeafletElement, group_1.LeafletGroup])
-    ], PolygonElement);
     return PolygonElement;
 }());
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", Array)
+], PolygonElement.prototype, "latlngs", void 0);
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", Object)
+], PolygonElement.prototype, "Options", void 0);
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", String)
+], PolygonElement.prototype, "mouseover", void 0);
+__decorate([
+    core_1.Input(),
+    __metadata("design:type", String)
+], PolygonElement.prototype, "onclick", void 0);
+PolygonElement = __decorate([
+    core_1.Component({
+        moduleId: module.id,
+        selector: 'polygon-element',
+        templateUrl: 'polygon.html',
+        styleUrls: ['polygon.css']
+    }),
+    __param(5, core_1.Optional()),
+    __param(6, core_1.Optional()),
+    __metadata("design:paramtypes", [map_service_1.MapService,
+        group_service_1.GroupService,
+        popup_service_1.PopupService,
+        globalId_service_1.GuidService,
+        helper_service_1.HelperService,
+        map_1.LeafletElement,
+        group_1.LeafletGroup])
+], PolygonElement);
 exports.PolygonElement = PolygonElement;
 //# sourceMappingURL=polygon.js.map
