@@ -8,7 +8,7 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
 var __metadata = (this && this.__metadata) || function (k, v) {
     if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
 };
-var core_1 = require('@angular/core');
+var core_1 = require("@angular/core");
 var HelperService = (function () {
     function HelperService() {
     }
@@ -28,11 +28,11 @@ var HelperService = (function () {
         }
         return true;
     };
-    HelperService = __decorate([
-        core_1.Injectable(), 
-        __metadata('design:paramtypes', [])
-    ], HelperService);
     return HelperService;
 }());
+HelperService = __decorate([
+    core_1.Injectable(),
+    __metadata("design:paramtypes", [])
+], HelperService);
 exports.HelperService = HelperService;
 //# sourceMappingURL=helper.service.js.map
