@@ -8,7 +8,7 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
 var __metadata = (this && this.__metadata) || function (k, v) {
     if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
 };
-var core_1 = require('@angular/core');
+var core_1 = require("@angular/core");
 var PopupService = (function () {
     function PopupService() {
     }
@@ -32,11 +32,11 @@ var PopupService = (function () {
             });
         }
     };
-    PopupService = __decorate([
-        core_1.Injectable(), 
-        __metadata('design:paramtypes', [])
-    ], PopupService);
     return PopupService;
 }());
+PopupService = __decorate([
+    core_1.Injectable(),
+    __metadata("design:paramtypes", [])
+], PopupService);
 exports.PopupService = PopupService;
 //# sourceMappingURL=popup.service.js.map
